Histogram function ready, 5.2 is finished?
So just finished the implementation of the histogram function. I made the mistake in the thresholding segmentation that I reproduced the histogram of the noisy image with the different histogrms of the three labels. Of course that is true: in the thresholded/labeled image no pixel under 100 will be labeled other than label 1 so of course we don't see three nice curves. Hope that makes sense.