Added graph cuts for image segmentation
I'm importing the image and rescaling it from uint16 to float numbers 0.0 to 1.0. 
In the next step each pixel is assigned to a node in the graph with an integer label. Computing maximum flow through image graphs leads to a reasonable segmentation of bone and background for edge potentials between 0.4 and 0.6. (In the Matlab demo she showed, this is the "beta" parameter she played around with to get good results. )