Clean wrap up
In an attempt to unclutter the notebook for a more compact "presentation", I've wrapped all the function definitions in a helper_functions.py file that is run at the beginning of the notebook. 

I've also condensed our progress today into the key components. Next, we'll need to figure out how to initialize the circles as central as possible.